feat: add blob count checks (#4447)

+++ crates/primitives/src/constants/eip4844.rs
@@ -20,7 +20,7 @@ pub const MAX_DATA_GAS_PER_BLOCK: u64 = 786_432u64; // 0xC0000
 pub const TARGET_DATA_GAS_PER_BLOCK: u64 = 393_216u64; // 0x60000
 
 /// Maximum number of data blobs in a single block.
-pub const MAX_BLOBS_PER_BLOCK: u64 = MAX_DATA_GAS_PER_BLOCK / DATA_GAS_PER_BLOB; // 786432 / 131072  = 6
+pub const MAX_BLOBS_PER_BLOCK: usize = (MAX_DATA_GAS_PER_BLOCK / DATA_GAS_PER_BLOB) as usize; // 786432 / 131072  = 6
 
 /// Target number of data blobs in a single block.
 pub const TARGET_BLOBS_PER_BLOCK: u64 = TARGET_DATA_GAS_PER_BLOCK / DATA_GAS_PER_BLOB; // 393216 / 131072 = 3
